\name{phenotypes2genotypes-internal}

\alias{createMix}
\alias{plot4Nmixture}
\alias{createMix.test}

\alias{mapMarkers.internal}
\alias{ascendingMaptoJigSawMap}

\alias{genotypesToCross.internal}
\alias{writePhenotypes.internal}
\alias{lowerTrng.internal}
\alias{writeGenotypes.internal}
\alias{writeGenotypesMap.internal}
\alias{cleanNames.internal}
\alias{crossContainsMap.internal}
\alias{mergeChromosomes.internal}
\alias{majorityRule.internal}
\alias{doCleanUp.internal}

\alias{convertgenerate.biomarkers.internal}
\alias{splitPheno.internal}
\alias{filterRowSub.internal }
\alias{filterRow.internal}
\alias{sortMap.internal}

\alias{getChromosome.internal}
\alias{getYLocs.internal}
\alias{chromosomesLengths.internal}
\alias{makeChromPal.internal}
\alias{makeTransPal.internal}
\alias{switchChromosomes.internal}
\alias{removeChromosomes.internal}
\alias{removeChromosomesSub.internal}

\alias{splitPhenoRowEM.internal}
\alias{splitPhenoRow.internal}
\alias{print.population}
\alias{add.to.populationSubPheno.internal}
\alias{add.to.populationSubGeno.internal}
\alias{add.to.populationSubMap.internal}
\alias{fakeMixUps.internal}
\alias{getChrOffsets.internal}
\alias{getMarkerOffsetsFromMap}
\alias{getMarkerOffsets}
\alias{getPopulationOffsets.internal}
\alias{check.population}
\alias{fakePhysicalMap.internal}
\alias{cleanRfs.internal }


\alias{inRangeCheck.internal}
\alias{numericCheck.internal}
\alias{inListCheck.internal}
\alias{genotypeCheck.internal}

\alias{fakePheno.internal}
\alias{fakeFounders.internal}
\alias{convertMap.internal}

\alias{bestCorelated.internal}
\alias{checkParameters.internal}
\alias{simBC.internal}

\alias{checkBeforeOrdering}
\alias{findUsingTTest.internal}
\alias{add.to.populationSubGenoSub.internal}
\alias{add.to.populationSubPhenoSub.internal}
\alias{projectOldMarkers.internal}
\alias{recalculateMap.internal}
\alias{smooth.genoRow.internal}
\alias{smooth.genoSub.internal}

\alias{cross.denovo.internal}
\alias{rearrangeMarkers}
\alias{assignLinkageGroups}
\alias{map2mapCorrelationMatrix}
\alias{map2mapImage}
\alias{selectTopMarker.internal}
\alias{bestCorelatedSub.internal}
\alias{scoreMarker.internal}
\alias{getpeaks.internal}
\alias{QTLscan.internal}
\alias{bestQTL.internal}
\alias{defaultCheck.internal}
\alias{scoreMixups.internal}
\alias{generate.biomarkers.internal}
\alias{convertType.internal}

\alias{insertMarkers.internal}
\alias{printGeno.internal}

\alias{mergeInverted}
\alias{donothing}
\alias{msqrt}
\alias{mlog}
\alias{reciproce}
\alias{mprobit}
\alias{mlogit}
\alias{findrecombinations}
\alias{selectChr}
\alias{is.integer0}
\alias{read.populationHT.internal}
\alias{read.populationNormal.internal}
\alias{invertChromosome.internal}
\alias{load_line.internal}
\alias{load_linebyline.internal}
\alias{invertChromosome.internal}
\alias{t_test_line.internal}
\alias{t_test_linebyline.internal}

\title{Internal pheno2geno functions}

\description{
  Internal pheno2geno functions. These are generally not to be called by the user.
}

\usage{
  TODO
}


\author{ 
	Konrad Zych \email{k.zych@rug.nl}, Danny Arends \email{Danny.Arends@gmail.com}
	Maintainer: Konrad Zych \email{k.zych@rug.nl}
}

\keyword{internal}
