\name{testPopulation}
\alias{testPopulation}
\docType{data}

\title{Test population object}

\description{
  pheno2geno population object.
}  

\usage{data(testPopulation)}

\format{
  An object of class \code{\link{population}}. See \code{\link{create.population}} for details.
}



\keyword{datasets}