\name{phenotypes2genotypes-internal}

\alias{createMix}
\alias{analyseLineVariance}
\alias{applyFunctionToFile}
\alias{ascendingMaptoJigSawMap}
\alias{assignLinkageGroups}
\alias{checkAndBind}
\alias{checkForEpistasis}
\alias{cleanRfs}
\alias{donothing}
\alias{getMarkerOffsets}
\alias{getMarkerOffsetsFromMap}
\alias{matchMarkers}
\alias{mlog}
\alias{mlogit}
\alias{mprobit}
\alias{msqrt}
\alias{map2mapCorrelationMatrix}
\alias{map2mapImage}
\alias{mergeInverted}
\alias{normalModeReading}
\alias{plot4Nmixture}
\alias{print.population}
\alias{readSingleFile}
\alias{reciproce}
\alias{rearrangeMarkers}
\alias{selectByLine}
\alias{selectPhenotypes}
\alias{selectChr}
\alias{simulateParentalPhenotypes}
\alias{tTestByLine}
\alias{twoGenosModel}
\alias{writeSingleFile}

\alias{add.to.populationSubAnnots.internal}
\alias{add.to.populationSubGeno.internal}
\alias{add.to.populationSubMap.internal}
\alias{add.to.populationSubPheno.internal}

\alias{bestQTLSub.internal}
\alias{bestCorelated.internal}
\alias{bestCorelatedSub.internal}
\alias{bestQTL.internal}
\alias{chromosomesLengths.internal}
\alias{cleanGeno.internal}
\alias{cleanNames.internal}
\alias{convertMap.internal}
\alias{convertType.internal}
\alias{cross.denovo.internal}
\alias{doCleanUp.internal}
\alias{fakeFounders.internal}
\alias{fakeMixUps.internal}
\alias{fakePheno.internal}
\alias{fakePhysicalMap.internal}
\alias{filterRow.internal}
\alias{findUsingTTest.internal}
\alias{fullScanRow.internal}
\alias{generate.biomarkers.internal}
\alias{genotypeCheck.internal}
\alias{genotypesToCross.internal}
\alias{getChrOffsets.internal}
\alias{getChromosome.internal}
\alias{getMarkerOffsets.internal}
\alias{getPopulationOffsets.internal}
\alias{getpeaks.internal}
\alias{getYLocs.internal}
\alias{inListCheck.internal}
\alias{inRangeCheck.internal}
\alias{insertMarkers.internal}
\alias{invertChromosome.internal}
\alias{lowerTrng.internal}
\alias{makeChromPal.internal}
\alias{makeHaplos.internal}
\alias{mapMarkers.internal}
\alias{mergeChromosomes.internal}
\alias{mergeEnv.internal}
\alias{numericCheck.internal}
\alias{printGeno.internal}
\alias{removeChromosomesSub.internal}
\alias{scoreMixups.internal}
\alias{selectTopMarker.internal}
\alias{splitPheno.internal}
\alias{splitPhenoRowEM.internal}
\alias{simBC.internal}
\alias{switchChromosomes.internal}
\alias{writeGenotypes.internal}
\alias{writePhenotypes.internal}
\alias{selectByLineApply}
\alias{tTestByLineApply}

\alias{check.population}
\alias{defaultCheck.internal}
\alias{createMix.test}
\alias{is.integer0}

\alias{processInteractions}
\alias{splitPheno.Apply}

\title{Internal pheno2geno functions}

\description{
  Internal pheno2geno functions. These are generally not to be called by the user.
}

\usage{
  TODO
}


\author{ 
	Konrad Zych \email{k.zych@rug.nl}, Danny Arends \email{Danny.Arends@gmail.com}
	Maintainer: Konrad Zych \email{k.zych@rug.nl}
}

\keyword{internal}
